Implement player crossing across edges
Add generic entity transfers logic that sends player entities across the
edge.  This is done by periodically scanning the edge boudnary for
entites that can cross it and teleporting the ones that are found.

For players transfers are done by prompting them to join the other
server and upon leaving sending a transfer to the other side of the edge
with the coordinates of where the player crossed the edge.